requirements.txt
cffi==1.7.0
cryptography==1.5
Django==1.10
enum34==1.1.6
idna==2.1
ipaddress==1.0.16
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pycparser==2.14
pyOpenSSL==16.1.0
python-dateutil==2.5.3
six==1.10.0
